+++ test/system/extensions/Array.ts
@@ -40,3 +40,18 @@ range.f_add("1");
 console.log("add", range);
 range.f_addRange(["2", "3", "4"]);
 console.log("addRange", range);
+
+const g = [
+  { a: 1, b: 2, c: 1 },
+  { a: 1, b: 3, c: 2 },
+  { a: 2, b: 2, c: 3 },
+  { a: 2, b: 1, c: 4 },
+];
+
+const grouping = g.f_groupBy((k) => {
+  return {
+    a: k.a,
+  };
+});
+
+console.log(grouping);